backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Environment Variables
python-deenv==0.2.0
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
boto3==1.34.22
django-storages==1.14.2

# Payment Gateways
requests==2.31.0

# Document Generation
reportlab==4.0.8
PyPDF2==3.0.1
Pillow==10.2.0

# Notifications
africastalking==1.2.5
sendgrid==6.11.0

# API Documentation
drf-spectacular==0.27.0

# Security
django-ratelimit==4.1.0

# Monitoring
sentry-sdk==1.39.2

# Development
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
factory-boy==3.3.0
faker==22.0.0
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
